126 | Including recipe README.md in GitHub Pages
Closes #126

Tweaked the titles of `README.md` under `recipes/` as these are now copied to `docs/` as part of the process of building
GitHub Pages using Sphinx. To achieve this I've moved from using a Resructued Text `index.rst` to a
[MyST](https://myst-parser.readthedocs.io/en/latest/index.html) (Markedly Structured Text) `index.md` and introduced the
`recipes.md` which indexes the copied files.

Would have been nice to include the original files directly but its challenging to include documents from outside of the
Sphinx directory. I found it possible to `include` documents but not a way to link to them as documents themselves,
hence hitting on the solution of copying them over. To achieve this I use `rsync` but only want to grab the `.md` files
and so include the `recipe_rsync_exclude.txt` which is a list of file types passed to `rsync` that are to be excluded
from this copy action.