versioneer > setuptools_scm; setup.{cfg,py} > pyproject.toml
Closes #186

Switches from versioneer to setuptools_scm for dynamic derivation of version based on Git Tags and if ahead of a tag the
Git commit.

Switches wholesale to using pyproject.toml for configuration. This introduces the need for Flake8-pyproject as a
dependency as flake8 is yet to switch to using this file for configuration (and is unlikely to do so).