Fix to centre_freq shift logic in HASPI code
The bug in the original HASPI code was being preserved by
enforcing it in the low level centre_freq function. This has
been changed so that centre_freq can now work as expected
and the bug has been lifted up to the top level HASPI function
where it can be clearly seen.