Fix to the CEC3 baseline recipes
The amplifier was appearing in the 'evaluate' stage, which was how
things were done for ICASSP 2023 where teams were not expected to
change it. It has now been moved into the 'enhance' stage so that
it can be changed by the teams.

Note, this has not changed the scores. The enhance + evaluate
still perform the same operations as before.

ie, enhance + (amplify + score) => (enhance + amplify) + score